docs: update docs and move to plantuml refactor: repackage
