+++ images/high_level_plugin_class_diagram.iuml
@@ -1,0 +1,91 @@
+@startuml
+
+skinparam {
+    shadowing false
+    CircledCharacterFontSize 8
+    ClassBackgroundColor #FFF
+    ClassArrowColor #666
+    ClassBorderColor #666
+    PackageBorderColor #AAA
+    ClassBorderThickness 2
+    PackageBackgroundColor #EEE
+    PackageTitleAlignment left
+    PackageFontSize 12
+    ClassFontStyle bold
+    ClassFontSize 12
+    ArrowFontSize 12
+    ArrowThickness 2
+    Padding 5
+    linetype polyline
+    NoteBorderColor #666
+    NoteBackgroundColor #FFF
+}
+
+hide empty members
+hide circle
+
+together {
+    package esm-dynamic-plugins <<Rectangle>> {
+        together {
+            class PluginManager
+            class Plugin
+            class ExtensionDescriptor
+        }
+        class ExtensionFactory
+    }
+
+    package cli-framework <<Rectangle>> {
+        class SubCommand
+    }
+}
+
+package plugin-api <<Rectangle>> {
+    class PluginDetails
+    class FlowscripterPlugin extends Plugin
+
+    together {
+        class FlowscripterExtensionDescriptor extends ExtensionDescriptor
+        class FlowscripterExtensionDetails
+    class CommandExtensionDescriptor
+    class ComponentExtensionDescriptor
+    }
+
+    class CommandExtensionDetails extends FlowscripterExtensionDetails
+    class ComponentExtensionDetails extends FlowscripterExtensionDetails
+
+    class FlowscripterExtensionFactory extends ExtensionFactory
+    class CommandExtensionFactory extends FlowscripterExtensionFactory
+    class ComponentExtensionFactory extends FlowscripterExtensionFactory
+
+    class FlowscripterCommand
+    class FlowscripterComponent
+    class RuntimeContext
+    class Runtime
+}
+
+FlowscripterCommand --|> SubCommand
+
+FlowscripterExtensionDescriptor <|-down- CommandExtensionDescriptor
+FlowscripterExtensionDescriptor <|-down- ComponentExtensionDescriptor
+
+PluginManager -right-> "*" Plugin : "registers"
+Plugin -right-> "1..*" ExtensionDescriptor
+ExtensionDescriptor -right-> ExtensionFactory
+FlowscripterPlugin --> PluginDetails : "pluginData"
+FlowscripterExtensionDescriptor --> FlowscripterExtensionDetails : "extensionData"
+FlowscripterCommand --> Runtime
+FlowscripterComponent --> RuntimeContext
+ComponentExtensionFactory ..> FlowscripterComponent : "creates"
+CommandExtensionFactory ..> FlowscripterCommand : "creates"
+ComponentExtensionFactory --> RuntimeContext
+CommandExtensionFactory --> Runtime
+
+PluginDetails : version
+PluginDetails : vendor details
+PluginDetails : plugin dependencies
+CommandExtensionDescriptor : COMMAND_EXTENSION_POINT_ID
+ComponentExtensionDescriptor : COMPONENT_EXTENSION_POINT_ID
+CommandExtensionDetails : arguments
+ComponentExtensionDetails : typed config
+
+@enduml
