+++ images/high_level_runtime_class_diagram.iuml
@@ -1,0 +1,66 @@
+@startuml
+
+skinparam {
+    shadowing false
+    CircledCharacterFontSize 8
+    ClassBackgroundColor #FFF
+    ClassArrowColor #666
+    ClassBorderColor #666
+    PackageBorderColor #AAA
+    ClassBorderThickness 2
+    PackageBackgroundColor #EEE
+    PackageTitleAlignment left
+    PackageFontSize 12
+    ClassFontStyle bold
+    ClassFontSize 12
+    ArrowFontSize 12
+    ArrowThickness 2
+    Padding 5
+    linetype polyline
+}
+
+hide empty members
+hide circle
+
+package plugin-api <<Rectangle>> {
+    class FlowscripterComponent
+    class FlowscripterCommand
+}
+
+package runtime-api <<Rectangle>> {
+    class Runtime
+    class RuntimeContext
+
+    class Registry
+
+    class SessionManager
+    class Session
+
+    class DatastoreManager
+    class Datastore extends FlowscripterComponent
+
+    class DataModel extends FlowscripterComponent
+    class DataEntry
+}
+
+FlowscripterComponent ..> RuntimeContext: can access
+FlowscripterCommand ..> Runtime: can access
+
+RuntimeContext --> Registry
+RuntimeContext --> SessionManager
+RuntimeContext --> DatastoreManager
+
+Registry --> "*" Datastore
+Registry --> "*" DataModel
+
+SessionManager --> "*" Session : manages
+
+DatastoreManager --> "           *" Datastore : manages
+
+Datastore -right-> "*" DataEntry: stores
+
+Session --> "             *" DataEntry: contains
+
+DataEntry -> DataModel: defined by
+
+@enduml
