--- src/context/datastore/Datastore.ts
@@ -1,11 +1,0 @@
-/**
- * @module @flowscripter/plugin-api
- */
-
-import FlowscripterComponent from '../../component/FlowscripterComponent';
-
-/**
- * Interface exposed by a Flowscripter datastore implementation.
- */
-export default interface Datastore extends FlowscripterComponent {
-}
