--- src/context/session/Session.ts
@@ -1,9 +1,0 @@
-/**
- * @module @flowscripter/plugin-api
- */
-
-/**
- * Interface exposed by a Flowscripter session implementation.
- */
-export default interface Session {
-}
