+++ src/data/DataModel.ts
@@ -1,0 +1,14 @@
+/**
+ * @module @flowscripter/plugin-api
+ */
+
+/* eslint-disable @typescript-eslint/no-empty-interface */
+
+import FlowscripterComponent from '../component/FlowscripterComponent';
+
+/**
+ * Interface exposed by a Flowscripter data model implementation.
+ */
+export default interface DataModel extends FlowscripterComponent {
+
+}
