+++ src/session/SessionManager.ts
